Created a flexible CLI generator
This script contains a class for the creation of simple, but flexible,
CLIs.  It works by generating a “call” function on the fly, which is
named the name of the program.  In essence, it creates a method for
calling that function.

Also, command line options can be added, and toggled at will.  Thus, a
single CLI object can handle any number of programs, each with unique
command line options.